LPS-136117 Move FieldUtil to data-engine-js-components-web
- https://issues.liferay.com/browse/LPS-136117
