+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/core/reducers/fieldEditableReducer.es.js
@@ -12,16 +12,16 @@
  * details.
  */
 
-import {
-	FieldSupport,
-	FieldUtil,
-	SettingsContext,
-} from 'dynamic-data-mapping-form-builder';
+import {FieldSupport, SettingsContext} from 'dynamic-data-mapping-form-builder';
 
 import * as FormSupport from '../../utils/FormSupport.es';
 import {formatRules} from '../../utils/rulesSupport';
 import {PagesVisitor} from '../../utils/visitors.es';
 import {EVENT_TYPES} from '../actions/eventTypes.es';
+import {
+	createDuplicatedField,
+	findInvalidFieldReference,
+} from '../utils/fields';
 import {updateRulesReferences} from '../utils/rules';
 import sectionAdded from '../utils/sectionAddedHandler';
 import {enableSubmitButton} from '../utils/submitButtonController.es';
@@ -81,11 +81,7 @@ const updateFieldProperty = ({
 	) {
 		focusedField = SettingsContext.updateFieldReference(
 			focusedField,
-			FieldUtil.findInvalidFieldReference(
-				focusedField,
-				pages,
-				propertyValue
-			),
+			findInvalidFieldReference(focusedField, pages, propertyValue),
 			false
 		);
 	}
@@ -184,7 +180,7 @@ export default (state, action, config) => {
 				Object.keys(focusedField).length &&
 				propertyName === 'fieldReference' &&
 				(propertyValue === '' ||
-					FieldUtil.findInvalidFieldReference(
+					findInvalidFieldReference(
 						focusedField,
 						state.pages,
 						propertyValue
@@ -368,7 +364,7 @@ export default (state, action, config) => {
 				)
 			);
 
-			const newField = FieldUtil.createDuplicatedField(originalField, {
+			const newField = createDuplicatedField(originalField, {
 				availableLanguageIds,
 				defaultLanguageId,
 				editingLanguageId,
