+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/util/settingsContext.es.js
@@ -17,9 +17,9 @@ import {
 	generateInstanceId,
 	normalizeFieldName,
 } from 'data-engine-js-components-web';
+import {updateFieldValidationProperty} from 'data-engine-js-components-web/js/core/utils/fields';
 
 import {getDefaultFieldName} from '../../../util/fieldSupport.es';
-import {updateFieldValidationProperty} from './fields.es';
 
 export const getSettingsContextProperty = (
 	settingsContext,
