+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/index.es.js
@@ -12,7 +12,6 @@
  * details.
  */
 
-export * as FieldUtil from './components/LayoutProvider/util/fields.es';
 export * as SettingsContext from './components/LayoutProvider/util/settingsContext.es';
 export {Token} from './expressions/Token.es';
 export {Tokenizer} from './expressions/Tokenizer.es';
